+++ src/test/groovy/fr/nereide/test/inspection/XmlInspectionTest.groovy
@@ -1,0 +1,10 @@
+package fr.nereide.test.inspection
+
+class XmlInspectionTest extends BaseInspectionTest {
+
+    String getLang() { return 'xml' }
+
+    void testNoGroovyServiceFileFoundInspection() {
+        doTest()
+    }
+}