+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unrealeased]
+
+### Fixed
+- Fixed the recuperation of component dirs that brought back test directories and broke some references
+
 ## [1.5.0] Entity fields completion update
 
 ### Added